requirements.txt
torch>=2.0.0
transformers>=4.30.0
tqdm>=4.60.0
numpy>=1.20.0
faiss-cpu>=1.7.0
